pyproject.toml
[tool.poetry]
name = "proplens ai"
version = "0.1.0"
description = ""
authors = ["Rahul Pandey <rahulrpandey1105@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.2.14"
langgraph = "^0.2.4"
langchain-community = "^0.2.12"
langchain-ollama = "^0.1.1"
pypdf = "^4.3.1"
docx2txt = "^0.8"
sentence-transformers = "^3.0.1"
chromadb = "^0.5.5"
langchain-chroma = "^0.1.2"
langchain-huggingface = "^0.0.3"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
faiss-cpu = "^1.8.0.post1"
pytesseract = "^0.3.13"
pdfplumber = "^0.11.4"
pymupdf = "^1.24.9"
python-docx = "^1.1.2"
bs4 = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
